+++ client/src/app/layout.tsx
@@ -3,8 +3,7 @@ import { theme } from '@/theme/override';
 import { ConfigProvider, Layout } from 'antd';
 import type { Metadata } from 'next';
 import { Navbar } from '@/components/navbar';
-import { HeaderComponent } from '@/components/header';
-import { ContentWrapper } from '@/components/content';
+import { ContentWrapper } from '@/components/';
 
 import 'antd/dist/reset.css';
 import './globals.css';
@@ -22,10 +21,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
           <ConfigProvider theme={theme}>
             <Layout>
               <Navbar />
-              <ContentWrapper>
-                <HeaderComponent />
-                {children}
-              </ContentWrapper>
+              <ContentWrapper>{children}</ContentWrapper>
             </Layout>
           </ConfigProvider>
         </StyledComponentsRegistry>