+++ client/src/components/content.tsx
@@ -1,9 +1,14 @@
 'use client';
 import React, { PropsWithChildren } from 'react';
-import { Layout } from 'antd';
+import { Layout, Grid } from 'antd';
 
 const { Content } = Layout;
+const { useBreakpoint } = Grid;
 
 export const ContentWrapper: React.FC<PropsWithChildren> = ({ children }) => {
-  return <Content style={{ padding: '24px 0' }}>{children}</Content>;
+  const screens = useBreakpoint();
+
+  return (
+    <Content style={{ padding: screens.lg ? 48 : '24px 0', margin: screens.xl ? '0 36px' : 0 }}>{children}</Content>
+  );
 };