+++ client/src/theme/styles.ts
@@ -1,3 +1,5 @@
+import { Raleway, Source_Sans_3 } from 'next/font/google';
+
 export const headingLink: React.CSSProperties = {
   fontSize: 11,
   textTransform: 'uppercase',
@@ -5,5 +7,15 @@ export const headingLink: React.CSSProperties = {
   fontWeight: 800,
 };
 
-export const colorBorder = 'rgba(160, 160, 160, 0.3)';
 export const colorTextSecondary = '#646464';
+export const colorBorder = 'rgba(160, 160, 160, 0.3)';
+
+export const secondaryLink: React.CSSProperties = {
+  color: colorTextSecondary,
+  textTransform: 'uppercase',
+  letterSpacing: '0.25em',
+  borderBottom: `1px dotted ${colorBorder}`,
+};
+
+export const mainFont = Raleway({ subsets: ['latin', 'cyrillic-ext'] });
+export const secondaryFont = Source_Sans_3({ subsets: ['latin', 'cyrillic-ext'] });